# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MissionLedger API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with company code and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Invalid credentials
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Organization registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: Get chart of accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    put:
      operationId: updateAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /bank-accounts:
    get:
      operationId: getBankAccounts
      tags: [bank-accounts]
      summary: Get all bank accounts
      responses:
        "200":
          description: List of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bank-accounts]
      summary: Create a bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankAccountRequest"
      responses:
        "201":
          description: Bank account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}:
    put:
      operationId: updateBankAccount
      tags: [bank-accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankAccountRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    delete:
      operationId: deleteBankAccount
      tags: [bank-accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /bank-transactions:
    get:
      operationId: getBankTransactions
      tags: [bank-transactions]
      summary: Get bank transactions
      parameters:
        - name: bankAccountId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bank transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankTransaction"
    post:
      operationId: createBankTransaction
      tags: [bank-transactions]
      summary: Create a bank transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankTransactionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"
  /bank-transactions/{id}:
    put:
      operationId: updateBankTransaction
      tags: [bank-transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBankTransactionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"
    delete:
      operationId: deleteBankTransaction
      tags: [bank-transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /bank-transactions/{id}/categorize:
    post:
      operationId: categorizeBankTransaction
      tags: [bank-transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategorizeBankTransactionRequest"
      responses:
        "200":
          description: Categorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"
...
components:
  schemas:
...
    UserSession:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        companyId:
          type: string
        companyName:
          type: string
        companyCode:
          type: string
        organizationType:
          type: string
      required: [id, email, role, companyId]
    LoginRequest:
      type: object
      properties:
        companyCode:
          type: string
        email:
          type: string
        password:
          type: string
      required: [companyCode, email, password]
    RegisterRequest:
      type: object
      properties:
        organizationName:
          type: string
        ein:
          type: string
        organizationType:
          type: string
        adminName:
          type: string
        adminEmail:
          type: string
        password:
          type: string
      required: [organizationName, ein, organizationType, adminName, adminEmail, password]
...
    Fund:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        balance:
          type: number
        totalDonations:
          type: number
        totalExpenses:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, isActive, createdAt, updatedAt]
...
    Account:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        parentId:
          type: string
          nullable: true
        balance:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, code, name, type, isActive, createdAt, updatedAt]
    CreateAccountRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        parentId:
          type: string
      required: [code, name, type]
...
    BankAccount:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        accountType:
          type: string
        lastFour:
          type: string
          nullable: true
        currentBalance:
          type: number
        glAccountId:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, accountType, currentBalance, isActive, createdAt, updatedAt]
    CreateBankAccountRequest:
      type: object
      properties:
        name:
          type: string
        accountType:
          type: string
        lastFour:
          type: string
        currentBalance:
          type: number
        glAccountId:
          type: string
        isActive:
          type: boolean
      required: [name, accountType, currentBalance]
    BankTransaction:
      type: object
      properties:
        id:
          type: string
        bankAccountId:
          type: string
        date:
          type: string
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        amount:
          type: number
        type:
          type: string
        status:
          type: string
        fundId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        account:
          $ref: "#/components/schemas/Account"
          nullable: true
        fund:
          $ref: "#/components/schemas/Fund"
          nullable: true
        journalEntryId:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, bankAccountId, date, description, amount, type, status, createdAt, updatedAt]
    CreateBankTransactionRequest:
      type: object
      properties:
        bankAccountId:
          type: string
        date:
          type: string
        description:
          type: string
        merchantName:
          type: string
        amount:
          type: number
        type:
          type: string
      required: [bankAccountId, date, description, amount, type]
    UpdateBankTransactionRequest:
      type: object
      properties:
        description:
          type: string
        merchantName:
          type: string
        fundId:
          type: string
        accountId:
          type: string
        status:
          type: string
    CategorizeBankTransactionRequest:
      type: object
      properties:
        accountId:
          type: string
        fundId:
          type: string
        description:
          type: string
      required: [accountId]
...